requirements.txt
alembic==1.13.1 ; python_version >= "3.11" and python_version < "4.0"
aniso8601==9.0.1 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
black==24.2.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.7.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" 
flask-cors==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
flask-jwt-extended==4.6.0 ; python_version >= "3.11" and python_version < "4"
flask-migrate==4.0.7 ; python_version >= "3.11" and python_version < "4.0"
flask-restx==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
flask-sqlalchemy==3.1.1 ; python_version >= "3.11" and python_version < "4.0"
flask==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.17.1 ; python_version >= "3.11" and python_version < "4.0"
google-api-python-client==2.122.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth-oauthlib==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.28.2 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.63.0 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.11" and python_version < "4.0" 
gunicorn==21.2.0 ; python_version >= "3.11" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
importlib-resources==6.3.0 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.11" and python_version < "4.0"
mako==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.11" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.11" and python_version < "4"
pymysql==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.33.0 ; python_version >= "3.11" and python_version < "4.0"
requests-oauthlib==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
sqlalchemy==2.0.28 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.0.1 ; python_version >= "3.11" and python_version < "4"
